ACC/requirements.txt
streamlit==1.44.1
pandas==2.2.3
numpy==2.1.3
plotly==5.24.1
pyyaml==6.0.2
